function_app/requirements.txt
azure-functions
azure-identity
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-resource
